+++ test/integration/default/anaconda_install.rb
@@ -20,13 +20,13 @@ control 'jupyterhub-chef::anaconda_install' do
     its('group') { should eq 'root' }
   end
 
-  describe file('/opt/anaconda/downloads/Anaconda3-5.1.0-Linux-x86_64.sh') do
+  describe file('/opt/anaconda/downloads/Anaconda3-5.2.0-Linux-x86_64.sh') do
     it { should be_file }
     its('owner') { should eq 'root' }
     its('group') { should eq 'root' }
   end
 
-  describe file('/opt/anaconda/Anaconda3-5.1.0') do
+  describe file('/opt/anaconda/Anaconda3-5.2.0') do
     it { should be_directory }
     its('owner') { should eq 'root' }
     its('group') { should eq 'root' }